Move remaining logic into modules
Bug: T168455
