+++ editquality/codegen/util.py
@@ -1,0 +1,11 @@
+import collections
+
+
+# https://stackoverflow.com/a/3233356
+def deep_update(d, u):
+    for k, v in u.items():
+        if isinstance(v, collections.Mapping):
+            d[k] = deep_update(d.get(k, {}), v)
+        else:
+            d[k] = v
+    return d
